ONS Boundaries

Metadata about a subset of the geographical boundaries from ONS' geoportal that can be downloaded by {ukgeographies}. These boundaries have been retrieved from the services provided by ONS' geoportal API and automatically categorised from their URLS, to match Geoportal's menus and datasets.

Usage

ons_boundaries

Format

A data frame with 513 rows and 7 columns:

id

A unique identifier for every boundary

service

URL pointing to the API service

boundary_type

Type of boundary, according to ONS' classification

boundary

Boundary (short) name, according to ONS' naming

detail_level

Boundary's level of detail. For a detailed description of the methodology refer to Digital boundaries

year

Year in which the boundaries were created

url_download

URL querying the API service to return all features as a geojson file

Source

https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/